requirements.txt
opencv-python==4.10.0.84
numpy==1.26.4
Flask==3.0.3
requests==2.32.2
pillow==10.4.0
python-dotenv==1.0.1
boto3==1.35.34
keras==3.3.3
tensorflow==2.16.1
psycopg2-binary==2.9.10
flask_bcrypt==1.0.1
flask_jwt_extended==4.7.1
